소드맨 추가 및 몬스터 로직 일부 수정

+++ Engine_SOURCE/yaNavMesh.cpp
@@ -18,7 +18,7 @@ namespace ya
 	void NavMesh::Initialize()
 	{
 		mbNavOn = false;
-		mDestPos = Vector3(100.532, -23.113, -3.576);
+		mDestPos = Vector3((float)100.532, (float)-23.113, (float)-3.576);
 
 		mbTraceOn = true;
 		mSpeed = 80;
